requirements.txt
transformers
torch
torchvision
datasets
pillow
streamlit  # (optionnel)
